feat: Add employee data to database
Add employee data to the `db.json` file in the `client/employee-management` directory. This includes information such as employee names, job roles, salaries, and registration numbers. The data is structured as an array of objects, with each object representing an employee. This commit adds the initial set of employee data.

Also, update the `package.json` file to include a new script called "mock-api". This script uses JSON Server to watch the `db.json` file and serves it as a mock API on port 3000. This will be useful for testing and development purposes.